coverage: Add command for reading counter value
Facilitate checking coverage counters from scripts and tests with a new
"coverage/read-counter" command that gets the total count for a counter.

Same could be achieved by scraping the output of "coverage/show" command
but the difficulties there are that output is in human readable format
and zero-value counters are not listed.

Signed-off-by: Jakub Sitnicki <jkbs@redhat.com>