ovn-nbctl: Initial support for daemon mode.
Make ovn-nbctl act as a unixctl server if we were asked to detach. This
turns ovn-nbctl into a long-lived process that acts a proxy for
interacting with NB DB. The main difference to regular mode of ovn-nbctl
is that in the daemon mode, a local copy of database contents has to be
obtained only once.

Just two unixctl commands are supported 'run' and 'exit'. The former can
be used to run any ovn-nbctl command or a batch of them as so:

  ovs-appctl -t ovn-nbctl run [OPTIONS] COMMAND [-- [OPTIONS] COMMAND] ...

Commands that use tabular output ('find' and 'list') are not
supported. As are --dry-run, --timeout, and --wait ovn-nbctl options.

Also, running commands that have not yet been converted to not use
ctl_fatal() will result in death of the daemon process. However,
--monitor option can be used to keep the daemon running.

Signed-off-by: Jakub Sitnicki <jkbs@redhat.com>