ovsdb-idl: Allow monitoring a columns if it is already is monitored
If we have created the IDL with monitoring and alters turned on by
default for all columns, there is no harm in allowing the user to
ask again for monitoring and alters to be enabled.

This allows us to run prerequisites handlers for db-ctl commands once
the IDL has already ran.

Signed-off-by: Jakub Sitnicki <jkbs@redhat.com>